fix(guard): validate trust CLI backend probes
